EV-49 rewrite EventSchedule list, fix imports, rename all components file
