+++ src/components/EventSchedules/EventScheduleItemView.js
@@ -7,9 +7,9 @@ import { connect } from "react-redux";
 import Moment from "react-moment";
 import PropTypes from "prop-types";
 import "moment-timezone";
-import renderPeriod from "./render-period";
-import useStyles from "./card-style-const";
-import SelectiveForm from "./selective-form";
+import renderPeriod from "./RenderPeriod";
+import useStyles from "./cardStyleConst";
+import SelectiveForm from "./SelectiveForm";
 import "../layout/colorlib.css";
 import getEvent from "../../actions/event/event-item-view-action";
 import eventDefaultImage from "../../constants/eventDefaultImage";
