+++ src/components/EventSchedules/RenderPeriod.js
