+++ src/containers/cancel-all-events.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import { connect } from "react-redux";
 import PropTypes from "prop-types";
 import Dropdown from "react-bootstrap/Dropdown";
-import EventScheduleModal from "../components/eventSchedule/eventSchedule-modal";
+import EventScheduleModal from "../components/EventSchedules/EventScheduleModal";
 import cancelAllEventSchedules from "../actions/eventSchedule/eventSchedule-cancel-all-action";
 
 class CancelAllEventsWrapper extends Component {
